// src/autogen/interfaces/anthology/2022/02/26/SchemaTaggedPayload.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SchemaTaggedPayload {
  accessControlPolicy?: string;
  applicationVersion?: string | number;
  batch?: string;
  createdAt?: string | number;
  creator?: string;
  data: any;
  device?: string;
  fingerprint?: string;
  id?: string | number;
  isPreprocessed?: boolean;
  name?: string;
  project?: string;
  protocolVersion: string | number;
  schemaName: string;
  schemaUrl?: string;
  schemaVersion: string | number;
  [k: string]: unknown;
}
export interface TypedSchemaTaggedPayload<T> extends SchemaTaggedPayload { data: T }